Add in setter and getter function to set and get tournament size and mutation size for demo purpose later
Default Tournament Size changed to 2 instead of 5 because it has higher chances of getting best fitness value.

Observation:
1. Population size increases, the better the result is (both the result and the number of generations it needs to compute the best fitness value).
2. The cycle of evolution that the population goes through increases, it often does not change the result much.
3. Tournament size increases, the lesser the number of generation it needs to get the end result (whether it is the best fitness value or the value that it sticks to) but it either get the best result extremely fast or get stuck in the not so good value forever,
ex:
Best Fitness Value = 0
It either gets this value within 10 generations or it gets stuck in the cycle saying 33.99 is the best fitness value.
4. Tournament size decreases, the larger the number of generations it needs to get its end result.
ex:
Best Fitness Value = 0
It might takes up to 90 generations to get this value 0. Even if it doesn't get 0, it will get value very close to 0, such as 1 and 4 unlike the previous setting.